requirements.txt
kafka-python==2.0.2
pydantic>=2.7.0
python-dotenv==1.0.0
requests==2.31.0
redis==5.0.1
praw==7.7.1
feedparser==6.0.11